Fixed some obvious bugs.

+++ src/com/ai/problems/min_set/records/SolutionObjective.java
@@ -4,5 +4,5 @@ package com.ai.problems.min_set.records;
  * @param solutionId Id of the solution in memory.
  * @param objectiveValue Objective value of the solution.
  */
-public record SolutionObjective(int solutionId, double objectiveValue) {
+public record SolutionObjective(int solutionId, int objectiveValue) {
 }
